finish game file
