--- lib/game_start.rb
@@ -1,13 +1,0 @@
-class Game
-    attr_reader :start
-    def initialize
-        
-    def start
-        p 'Welcome to War! (or Peace) This game will be played with 52 cards.
-        The players today are Megan and Aurora.
-        Type 'GO' to start the game!'
-        
-
-
-    end 
-end
